fix(compiler): split numeric stack analyzers
